rename the env var file and example, split fixtures between unit testing and functional testing
