+++ dbt-athena/.env.example
@@ -1,10 +1,10 @@
 DBT_TEST_ATHENA_S3_STAGING_DIR=
 DBT_TEST_ATHENA_S3_TMP_TABLE_DIR=
 DBT_TEST_ATHENA_REGION_NAME=
-DBT_TEST_ATHENA_THREADS=
-DBT_TEST_ATHENA_POLL_INTERVAL=
 DBT_TEST_ATHENA_DATABASE=
 DBT_TEST_ATHENA_SCHEMA=
 DBT_TEST_ATHENA_WORK_GROUP=
+DBT_TEST_ATHENA_THREADS=
+DBT_TEST_ATHENA_POLL_INTERVAL=
+DBT_TEST_ATHENA_NUM_RETRIES=
 DBT_TEST_ATHENA_AWS_PROFILE_NAME=
-DBT_TEST_ATHENA_SPARK_WORK_GROUP=
